Added code to try and kill fires slowly. The room now has a stack of fires. When the water is on it should kill one fire every second.
When fires are spawned in the FireSpawner or the Fire I look for the room they are in via the transform in the FireSpawner and the RayCast in the Fire. The fire is then added to the stack of that room. The room is also added to the fire so that if the water is on the fire does not spawn any new fires. 

There is a problem in the sometimes all the fires are killed immediately. I haven't worked out why this is.